Done CRUD, Work Both Web And Mobile Apps

+++ Backend/assets/js/index.js
@@ -1,47 +1,50 @@
-$("#sampis").submit(function (event) {
-  alert("Data Inserted Succesfully");
-});
+$(document).ready(() => {
+  console.log("ready");
 
-$("#tolo").submit((event) => {
-  event.preventDefault();
-
-  let unindexed_array = $("#tolo").serializeArray();
-  let data = {};
-
-  $.map(unindexed_array, function (n, i) {
-    data[n["name"]] = n["value"];
+  $("#sampis").submit(function (event) {
+    alert("Data Inserted Succesfully");
   });
 
-  console.table(data);
-  let request = {
-    url: `http://localhost:9999/api/users/${data.id}`,
-    method: "PUT",
-    data: data,
-  };
-  $.ajax(request).done((response) => {
-    console.log(response);
-    alert("Data Updated Succesfully");
-  });
-});
+  $("#tolo").submit((event) => {
+    event.preventDefault();
 
-console.log(window.location.pathname);
-console.log("SAMPIS");
-if (window.location.pathname == "/") {
-  let $ondelete = $(".table tbody td a.delete");
-  $ondelete.click((event) => {
-    console.table($ondelete.attr);
-    let id = $ondelete.attr("data-id");
+    let unindexed_array = $(event.target).serializeArray();
+    let data = {};
 
+    $.map(unindexed_array, function (n, i) {
+      data[n["name"]] = n["value"];
+    });
+
+    console.table(data);
     let request = {
-      url: `http://localhost:9999/api/users/${id}`,
-      method: "DELETE",
+      url: `http://localhost:9999/api/users/${data.id}`,
+      method: "PUT",
+      data: data,
     };
-    console.table(request);
-    if (confirm("GAs Delete ? ")) {
-      $.ajax(request).done((response) => {
-        alert("Data Deleted Succesfully");
-        location.reload();
-      });
-    }
+    $.ajax(request).done((response) => {
+      console.log(response);
+      alert("Data Updated Succesfully");
+    });
   });
-}
+
+  console.log(window.location.pathname);
+  if (window.location.pathname == "/") {
+    jQuery(".table tbody td a.delete").click((evt) => {
+      evt.preventDefault();
+      let id = jQuery($(evt.target)).attr("data-id");
+      if (id) {
+        let request = {
+          url: `http://localhost:9999/api/users/${id}`,
+          method: "DELETE",
+        };
+        console.table(request);
+        if (confirm("GAs Delete ? ")) {
+          jQuery.ajax(request).done((response) => {
+            alert("Data Deleted Succesfully");
+            location.reload();
+          });
+        }
+      }
+    });
+  }
+});
